refactor(api): move and rename get-user-details-for-admin controller
