+++ api/tests/identity-access-management/unit/application/user/user.admin.controller.test.js
@@ -1,3 +1,4 @@
+import { usecases as libUsecases } from '../../../../../lib/domain/usecases/index.js';
 import { userAdminController } from '../../../../../src/identity-access-management/application/user/user.admin.controller.js';
 import { User } from '../../../../../src/identity-access-management/domain/models/User.js';
 import { usecases } from '../../../../../src/identity-access-management/domain/usecases/index.js';
@@ -157,6 +158,31 @@ describe('Unit | Identity Access Management | Application | Controller | Admin |
 
       // then
       expect(response).to.be.equal(newEmail);
+    });
+  });
+
+  describe('#getUserDetails', function () {
+    let request;
+    let dependencies;
+
+    beforeEach(function () {
+      request = { params: { id: 123 } };
+
+      sinon.stub(libUsecases, 'getUserDetailsForAdmin');
+      const userDetailsForAdminSerializer = { serialize: sinon.stub() };
+      dependencies = { userDetailsForAdminSerializer };
+    });
+
+    it('gets the specified user', async function () {
+      // given
+      libUsecases.getUserDetailsForAdmin.withArgs({ userId: 123 }).resolves('userDetail');
+      dependencies.userDetailsForAdminSerializer.serialize.withArgs('userDetail').returns('ok');
+
+      // when
+      const response = await userAdminController.getUserDetails(request, hFake, dependencies);
+
+      // then
+      expect(response).to.be.equal('ok');
     });
   });
 });
